Support running on macOS 12
Alas, macOS doesn't have `/proc/uptime` to give a handy
"centiseconds" counter. Some twiddling and investigation
suggests that the `machdep.time_since_reset` sysctl is an
incrementing counter (albeit at a slightly odd 24MHz) that
measures the passing of time.

Since it's easier for the shell script launcher to find the OS,
move `timex()` into an include file (`lib/time-linux.awk`,
`lib/time-darwin.awk`) and add the correct one to the invocation.

(Annoyingly, this only seems to exist in macOS 12.  Or at least
it's not in 11.6.4.  But it's better than nothing...)